requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
google-api-python-client==2.143.0
requests==2.32.3
